+++ test/integration/src/segments/user/UserSegment.js
@@ -2,8 +2,12 @@ import MapSegment from 'soya/lib/data/redux/segment/map/MapSegment.js';
 import UserActionCreator from './UserActionCreator.js';
 
 export default class UserSegment extends MapSegment {
-  _getName() {
+  static get segmentName() {
     return 'user';
+  }
+
+  _getName() {
+    return UserSegment.segmentName;
   }
 
   _createActionCreator() {